# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PAY API",
...
        "version": "1.0.0"
    },
    "paths": {
        "/api/account/get-balance": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get balance",
                "operationId": "ebf20ff1877e9a911db7245f5782fcdb",
                "parameters": [
                    {
                        "name": "api-token",
                        "in": "header",
                        "description": "token allow create service",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sign": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "query": {
                                        "id": 1
                                    },
                                    "sign": "signiture data"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result_false": {
                                        "summary": "Result error.",
                                        "value": {
                                            "error_code": 401,
                                            "system_time": "2023-11-24 11:15:59",
                                            "message": "access_token hết hạn hoặc không được phép sử dụng",
                                            "errors": [
                                                "access_token hết hạn hoặc không được phép sử dụng"
                                            ],
                                            "data": null,
                                            "tranid_tracking": "bc47d534b97f7cdb4c2d5d229376915c"
                                        }
                                    },
                                    "result_success": {
                                        "summary": "Result success.",
                                        "value": {
                                            "error_code": 200,
                                            "message": "success",
                                            "system_time": "2023-11-24 11:15:59",
                                            "errors": [],
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/account/create-qr-payment": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Tạo QRCode Thanh toán cá nhân",
                "operationId": "fc10e08e01848acefac4f6d8ac4e4db1",
                "parameters": [
                    {
                        "name": "api-token",
                        "in": "header",
                        "description": "token allow create service",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number"
                                    },
                                    "remark": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "amount": 10000,
                                    "remark": "test"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result_false": {
                                        "summary": "Result error.",
                                        "value": {
                                            "error_code": 401,
                                            "system_time": "2023-11-24 11:15:59",
                                            "message": "access_token hết hạn hoặc không được phép sử dụng",
                                            "errors": [
                                                "access_token hết hạn hoặc không được phép sử dụng"
                                            ],
                                            "data": null,
                                            "tranid_tracking": "bc47d534b97f7cdb4c2d5d229376915c"
                                        }
                                    },
                                    "result_success": {
                                        "summary": "Result success.",
                                        "value": {
                                            "error_code": 200,
                                            "message": "success",
                                            "system_time": "2023-11-24 11:15:59",
                                            "errors": [],
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}